requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==5.0.1
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-extensions==3.2.3
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
executing==2.0.1
iniconfig==2.0.0
ipython==8.21.0
isoweek==1.3.3
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.0.0
python-dateutil==2.8.2
pytz==2024.1
screen==1.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
TimeConvert==3.0.13
traitlets==5.14.1
tzlocal==5.2
wcwidth==0.2.13
wheel==0.42.0
xlwt==1.3.0